@@ -2,13 +2,13 @@
 
   <PropertyGroup>
     <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Common.Logging.Core" Version="3.4.1" />
-    <PackageReference Include="Google.Protobuf" Version="3.11.4" />
+    <PackageReference Include="Google.Protobuf" Version="3.12.1" />
     <PackageReference Include="Nethereum.ABI" Version="3.7.1" />
     <PackageReference Include="Nethereum.Contracts" Version="3.7.1" />
     <PackageReference Include="Nethereum.Hex" Version="3.7.1" />
@@ -24,7 +24,14 @@
   </ItemGroup>
 
   <Target Name="PostBuild" AfterTargets="PostBuildEvent">
-    <Exec Command="copy $(TargetPath) $(ProjectDir)\..&#xD;&#xA;copy $(TargetDir)$(TargetName).pdb ..\..\.." />
+		<Copy SourceFiles="$(TargetPath)" 
+			  DestinationFolder="$(ProjectDir).." 
+			  OverwriteReadOnlyFiles="true" 
+			  SkipUnchangedFiles="false" />
+		<Copy SourceFiles="$(TargetDir)$(TargetName).pdb" 
+			  DestinationFolder="..\..\.." 
+			  OverwriteReadOnlyFiles="true" 
+			  SkipUnchangedFiles="false" />
   </Target>
 
 </Project>
