@@ -2,12 +2,16 @@
 
   <PropertyGroup>
     <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
   </PropertyGroup>
 
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
+  </PropertyGroup>
+
   <ItemGroup>
-    <PackageReference Include="NBitcoin" Version="5.0.20" />
+    <PackageReference Include="NBitcoin" Version="5.0.39" />
   </ItemGroup>
 
   <ItemGroup>
@@ -15,7 +19,10 @@
   </ItemGroup>
 
   <Target Name="PostBuild" AfterTargets="PostBuildEvent">
-    <Exec Command="copy $(TargetPath) $(ProjectDir).." />
+    <Copy SourceFiles="$(TargetPath)" 
+			  DestinationFolder="$(ProjectDir).." 
+			  OverwriteReadOnlyFiles="true" 
+			  SkipUnchangedFiles="false" />
   </Target>
 
 </Project>
