@@ -7,8 +7,8 @@ namespace ccaas.Models
 {
     public class OfferQuery
     {
-        public string askOrderId;
-        public string bidOrderId;
-        public string expiration;
+        public string askOrderId { get; set; }
+        public string bidOrderId { get; set; }
+        public string expiration { get; set; }
     }
 }
