@@ -7,8 +7,8 @@ namespace ccaas.Models
 {
     public class OrderResponse
     {
-        public string id;
-        public string block;
-        public OrderQuery order;
+        public string id { get; set; }
+        public string block { get; set; }
+        public OrderQuery order { get; set; }
     }
 }
