@@ -7,11 +7,11 @@ namespace ccaas.Models
 {
     public class TransferQuery
     {
-        public string ethKey;
-        public string gain;
-        public string dealOrderId;
-        public string txid;
-        public string fee;
-        public string continuation;
+        public string ethKey { get; set; }
+        public string gain { get; set; }
+        public string dealOrderId { get; set; }
+        public string txid { get; set; }
+        public string fee { get; set; }
+        public string continuation { get; set; }
     }
 }
