@@ -10,28 +10,28 @@ namespace ccbe.Models
     public class Block
     {
         /// <summary>The block's height</summary>
-        public string BlockNum;
+        public string BlockNum { get; set; }
         /// <summary>The ID of consensus used to mine the block</summary>
-        public string Consensus;
+        public string Consensus { get; set; }
         /// <summary>The difficulty used to mine the block</summary>
-        public string Difficulty;
+        public string Difficulty { get; set; }
         /// <summary>The nonce used to mine the block</summary>
-        public string Nonce;
+        public string Nonce { get; set; }
         /// <summary>The time when the block had been mined</summary>
-        public string Timestamp;
+        public string Timestamp { get; set; }
         /// <summary>The reward for mining the block</summary>
-        public string BlockReward;
+        public string BlockReward { get; set; }
         /// <summary>The version of Creditcoin transaction processor used to mine the block</summary>
-        public string Version;
+        public string Version { get; set; }
         /// <summary>An estimate of the block's size</summary>
-        public string Size;
+        public string Size { get; set; }
         /// <summary>The ID of the block immediately bellow in the blockchain</summary>
-        public string PrevBlockId;
+        public string PrevBlockId { get; set; }
         /// <summary>The public key of the miner who mined the block</summary>
-        public string SignerPubKey;
+        public string SignerPubKey { get; set; }
         /// <summary>The sighash of the miner who mined the block</summary>
-        public string Sighash;
+        public string Sighash { get; set; }
         /// <summary>A list of transactions in the block</summary>
-        public Dictionary<string, Transaction> Transactions;
+        public Dictionary<string, Transaction> Transactions { get; set; }
     }
 }
